requirements.txt
GitPython
python-dotenv
pytz
